Extract upgrade button sealed class
