+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/viewmodel/OnboardingUpgradeFeaturesViewModel.kt
@@ -4,18 +4,18 @@ import android.app.Activity
 import android.app.Application
 import android.content.Context
 import android.view.accessibility.AccessibilityManager
-import androidx.annotation.StringRes
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
 import au.com.shiftyjelly.pocketcasts.account.onboarding.upgrade.FeatureCardsState
+import au.com.shiftyjelly.pocketcasts.account.onboarding.upgrade.UpgradeButton
 import au.com.shiftyjelly.pocketcasts.account.onboarding.upgrade.UpgradeFeatureCard
+import au.com.shiftyjelly.pocketcasts.account.onboarding.upgrade.toUpgradeButton
 import au.com.shiftyjelly.pocketcasts.account.onboarding.upgrade.toUpgradeFeatureCard
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.models.type.RecurringSubscriptionPricingPhase
 import au.com.shiftyjelly.pocketcasts.models.type.Subscription
-import au.com.shiftyjelly.pocketcasts.models.type.Subscription.SubscriptionTier
 import au.com.shiftyjelly.pocketcasts.models.type.SubscriptionFrequency
 import au.com.shiftyjelly.pocketcasts.models.type.SubscriptionMapper
 import au.com.shiftyjelly.pocketcasts.models.type.SubscriptionPricingPhase
@@ -36,7 +36,6 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.reactive.asFlow
 import timber.log.Timber
 import javax.inject.Inject
-import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @HiltViewModel
 class OnboardingUpgradeFeaturesViewModel @Inject constructor(
@@ -268,38 +267,7 @@ sealed class OnboardingUpgradeFeaturesState {
     }
 }
 
-private fun Subscription.toUpgradeButton() = when (this.tier) {
-    SubscriptionTier.PLUS -> UpgradeButton.Plus(this)
-    SubscriptionTier.PATRON -> UpgradeButton.Patron(this)
-    SubscriptionTier.UNKNOWN -> throw IllegalStateException("Unknown subscription tier")
-}
-
 private fun RecurringSubscriptionPricingPhase.toSubscriptionFrequency() = when (this) {
     is SubscriptionPricingPhase.Months -> SubscriptionFrequency.MONTHLY
     is SubscriptionPricingPhase.Years -> SubscriptionFrequency.YEARLY
-}
-
-sealed class UpgradeButton(
-    @StringRes val shortNameRes: Int,
-    val backgroundColor: Long,
-    val textColor: Long,
-    open val subscription: Subscription,
-) {
-    data class Plus(
-        override val subscription: Subscription,
-    ) : UpgradeButton(
-        shortNameRes = LR.string.pocket_casts_plus_short,
-        backgroundColor = 0xFFFFD846,
-        textColor = 0xFF000000,
-        subscription = subscription,
-    )
-
-    data class Patron(
-        override val subscription: Subscription,
-    ) : UpgradeButton(
-        shortNameRes = LR.string.pocket_casts_patron_short,
-        backgroundColor = 0xFF6046F5,
-        textColor = 0xFFFFFFFF,
-        subscription = subscription,
-    )
 }
